fixed unit tests for new Donor class
